backend/requirements.txt
absl-py==0.11.0
aniso8601==8.1.0
argcomplete==1.10.0
astunparse==1.6.3
beautifulsoup4==4.8.0
bidict==0.21.2
blist==1.3.6
cachetools==4.2.0
cassandra-driver==3.24.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-CQLAlchemy==1.2.0
flask-redis==0.4.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.0.1
flatbuffers==1.12
gast==0.3.3
geomet==0.2.1.post1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.2
h5py==2.10.0
idna==2.10
IMAPClient==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
logstash-formatter==0.5.17
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pdfminer.six==20181108
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.9
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.0.0
python-pptx==0.6.18
python-socketio==5.0.3
pytz==2020.4
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.1
SpeechRecognition==3.8.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
textract==1.6.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.2
uWSGI==2.0.19.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7
